[lldb-dap][windows] add support for out of PATH python.dll resolution (#179306)
(cherry picked from commit 67e571df8c274f6e5357e822d79ca2c34c0056c4)
